Added required functions from noise_module.py into channelcatalog.py and deleted noise_module.py
