+++ pyproject.toml
@@ -15,13 +15,10 @@ classifiers = [
 ]
 dynamic = ["version"]
 dependencies = [
-    "ipykernel", # Support for Jupyter notebooks
     "DateTimeRange>=2.0.0,<3.0.0",
     "numpy>=1.22.0,<2.0.0",
     "pandas>=1.5.3,<2.0.0",
-    "numba>=0.57.0,<1.0.0",
     "pyasdf>=0.7.5,<1.0.0",
-    "pycwt>=0.3.0a22,<1.0.0",
     "fsspec>=2023.4.0,<2024.0.0",
     "s3fs==2023.4.0,<2024.0.0",
     "zarr==2.14.2",
